chore: Move aws credentials to plugins/common.
