+++ plugins/common/aws/credentials.go
